requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
bs4
unstructured
pypdb
wikipedia
faiss-cpu
fastapi
langserve
sse_starlette